+++ saleor/checkout/urls.py
@@ -1,15 +1,8 @@
 from django.conf.urls import url
 
 from . import views
-from saleor.cart.views.discount import remove_voucher_view
+from .views.discount import remove_voucher_view
 
-cart_urlpatterns = [
-    url(r'^$', views.cart_index, name='index'),
-    url(r'^update/(?P<variant_id>\d+)/$',
-        views.update_cart_line, name='update-line'),
-    url(r'^summary/$', views.cart_summary, name='summary'),
-    url(r'^shipping-options/$', views.cart_shipping_options,
-        name='shipping-options')]
 
 checkout_urlpatterns = [
     url(r'^$', views.checkout_index, name='index'),
@@ -21,3 +14,12 @@ checkout_urlpatterns = [
     url(r'^remove_voucher/', remove_voucher_view,
         name='remove-voucher'),
     url(r'^login/', views.checkout_login, name='login')]
+
+
+cart_urlpatterns = [
+    url(r'^$', views.cart_index, name='index'),
+    url(r'^update/(?P<variant_id>\d+)/$',
+        views.update_cart_line, name='update-line'),
+    url(r'^summary/$', views.cart_summary, name='summary'),
+    url(r'^shipping-options/$', views.cart_shipping_options,
+        name='shipping-options')]
